refactor: add selectors folder and containerSelectors for test files
